+++ merlin/models/tf/blocks/core/inputs.py
@@ -25,6 +25,8 @@ from ...features.continuous import ContinuousFeatures
 from ...features.embedding import (
 from merlin.models.tf.blocks.base import Block, BlockType, ParallelBlock, TabularAggregationType
 from merlin.models.tf.blocks.core.aggregation import SequenceAggregation, SequenceAggregator
+from merlin.models.tf.blocks.core.base import Block, BlockType
+from merlin.models.tf.blocks.core.combinators import ParallelBlock, TabularAggregationType
 from merlin.models.tf.blocks.core.masking import MaskingBlock, masking_registry
 from merlin.models.tf.features.continuous import ContinuousFeatures
 from merlin.models.tf.features.embedding import (
