+++ merlin/models/tf/blocks/core/masking.py
@@ -20,7 +20,7 @@ import tensorflow as tf
 from tensorflow.keras import backend
 from tensorflow.python.ops import array_ops
 
-from merlin.models.tf.blocks.base import Block, PredictionOutput
+from merlin.models.tf.blocks.core.base import Block, PredictionOutput
 from merlin.models.utils.doc_utils import docstring_parameter
 from merlin.models.utils.registry import Registry
 from merlin.schema import Tags
