+++ merlin/models/tf/utils/batch_utils.py
@@ -8,7 +8,7 @@ from merlin.core.dispatch import DataFrameType, concat_columns, get_lib
 from merlin.models.tf.blocks.base import Block
 from merlin.models.tf.dataset import BatchedDataset
 from merlin.models.tf.models.base import Model, RetrievalModel
-from merlin.models.utils.schema import select_targets
+from merlin.models.utils.schema_utils import select_targets
 from merlin.schema import Schema, Tags
